Moved code out of MetadataTable
