+++ Source/Scene/MetadataTable.js
@@ -1,12 +1,9 @@
 import Check from "../Core/Check.js";
-import ComponentDatatype from "../Core/ComponentDatatype.js";
 import defaultValue from "../Core/defaultValue.js";
 import defined from "../Core/defined.js";
 import DeveloperError from "../Core/DeveloperError.js";
-import FeatureDetection from "../Core/FeatureDetection.js";
-import getStringFromTypedArray from "../Core/getStringFromTypedArray.js";
-import oneTimeWarning from "../Core/oneTimeWarning.js";
 import MetadataEntity from "./MetadataEntity.js";
+import MetadataTableProperty from "./MetadataTableProperty.js";
 import MetadataType from "./MetadataType.js";
 
 /**
@@ -35,12 +32,12 @@ function MetadataTable(options) {
   if (defined(options.properties)) {
     for (var propertyId in options.properties) {
       if (options.properties.hasOwnProperty(propertyId)) {
-        properties[propertyId] = initializeProperty(
-          count,
-          options.properties[propertyId],
-          options.class.properties[propertyId],
-          options.bufferViews
-        );
+        properties[propertyId] = new MetadataTableProperty({
+          count: count,
+          property: options.properties[propertyId],
+          classProperty: options.class.properties[propertyId],
+          bufferViews: options.bufferViews,
+        });
       }
     }
   }
@@ -83,7 +80,7 @@ Object.defineProperties(MetadataTable.prototype, {
    * A dictionary containing properties.
    *
    * @memberof MetadataTable.prototype
-   * @type {Object}
+   * @type {Object.<String, MetadataTableProperty>}
    * @readonly
    * @private
    */
@@ -205,7 +202,11 @@ MetadataTable.prototype.setProperty = function (index, propertyId, value) {
   }
 
   //>>includeStart('debug', pragmas.debug);
-  checkType(value, property.classProperty);
+  var classProperty = this._class.properties[propertyId];
+  var errorMessage = classProperty.validate(value);
+  if (defined(errorMessage)) {
+    throw new DeveloperError(errorMessage);
+  }
   //>>includeEnd('debug');
 
   value = MetadataEntity.unnormalize(this._class, propertyId, value);
@@ -277,117 +278,6 @@ function getDefault(classDefinition, propertyId) {
   }
 }
 
-function throwTypeError(value, type) {
-  throw new DeveloperError("value " + value + " does not match type " + type);
-}
-
-function throwOutOfRangeError(value, type, normalized) {
-  var errorMessage = "value " + value + " is out of range for type " + type;
-  if (normalized) {
-    errorMessage += " (normalized)";
-  }
-  throw new DeveloperError(errorMessage);
-}
-
-function checkInRange(value, type, normalized) {
-  if (normalized) {
-    var min = MetadataType.isUnsignedIntegerType(type) ? 0.0 : -1.0;
-    var max = 1.0;
-    if (value < min || value > max) {
-      throwOutOfRangeError(value, type, normalized);
-    }
-    return;
-  }
-
-  if (
-    value < MetadataType.getMinimum(type) ||
-    value > MetadataType.getMaximum(type)
-  ) {
-    throwOutOfRangeError(value, type, normalized);
-  }
-}
-
-function checkValue(value, classProperty) {
-  var javascriptType = typeof value;
-
-  var enumType = classProperty.enumType;
-  if (defined(enumType)) {
-    if (javascriptType !== "string" || !defined(enumType.valuesByName[value])) {
-      throw new DeveloperError(
-        "value " + value + " is not a valid enum name for " + enumType.id
-      );
-    }
-    return;
-  }
-
-  var valueType = classProperty.valueType;
-  var normalized = classProperty.normalized;
-
-  switch (valueType) {
-    case MetadataType.INT8:
-    case MetadataType.UINT8:
-    case MetadataType.INT16:
-    case MetadataType.UINT16:
-    case MetadataType.INT32:
-    case MetadataType.UINT32:
-      if (javascriptType !== "number") {
-        throwTypeError(value, valueType);
-      }
-      checkInRange(value, valueType, normalized);
-      break;
-    case MetadataType.INT64:
-    case MetadataType.UINT64:
-      if (javascriptType !== "number" && javascriptType !== "bigint") {
-        throwTypeError(value, valueType);
-      }
-      checkInRange(value, valueType, normalized);
-      break;
-    case MetadataType.FLOAT32:
-      if (javascriptType !== "number") {
-        throwTypeError(value, valueType);
-      }
-      if (isFinite(value)) {
-        checkInRange(value, valueType, normalized);
-      }
-      break;
-    case MetadataType.FLOAT64:
-      if (javascriptType !== "number") {
-        throwTypeError(value, valueType);
-      }
-      break;
-    case MetadataType.BOOLEAN:
-      if (javascriptType !== "boolean") {
-        throwTypeError(value, valueType);
-      }
-      break;
-    case MetadataType.STRING:
-      if (javascriptType !== "string") {
-        throwTypeError(value, valueType);
-      }
-      break;
-  }
-}
-
-function checkType(value, classProperty) {
-  if (classProperty.type === MetadataType.ARRAY) {
-    if (!Array.isArray(value)) {
-      throwTypeError(value, classProperty.type);
-    }
-    var length = value.length;
-    if (
-      defined(classProperty.componentCount) &&
-      classProperty.componentCount !== length
-    ) {
-      throw new DeveloperError("Array length does not match componentCount");
-    }
-    for (var i = 0; i < length; ++i) {
-      checkValue(value[i], classProperty);
-    }
-  } else {
-    checkValue(value, classProperty);
-  }
-}
-
 function checkIndex(table, index) {
   var count = table._count;
   if (!defined(index) || index < 0 || index >= count) {
@@ -396,549 +286,6 @@ function checkIndex(table, index) {
       "index is required and between zero and count - 1 (" + maximumIndex + ")."
     );
   }
-}
-
-function initializeProperty(count, property, classProperty, bufferViews) {
-  var i;
-
-  var isArray = classProperty.type === MetadataType.ARRAY;
-  var isVariableSizeArray = isArray && !defined(classProperty.componentCount);
-
-  var valueType = classProperty.valueType;
-
-  var hasStrings = valueType === MetadataType.STRING;
-  var hasBooleans = valueType === MetadataType.BOOLEAN;
-
-  var offsetType = defaultValue(
-    MetadataType[property.offsetType],
-    MetadataType.UINT32
-  );
-
-  var arrayOffsets;
-  if (isVariableSizeArray) {
-    arrayOffsets = new BufferView(
-      bufferViews[property.arrayOffsetBufferView],
-      offsetType,
-      count + 1
-    );
-  }
-
-  var componentCount;
-  if (isVariableSizeArray) {
-    componentCount = 0;
-    for (i = 0; i < count; ++i) {
-      componentCount += arrayOffsets.get(i + 1) - arrayOffsets.get(i);
-    }
-  } else if (isArray) {
-    componentCount = count * classProperty.componentCount;
-  } else {
-    componentCount = count;
-  }
-
-  var stringOffsets;
-  if (hasStrings) {
-    stringOffsets = new BufferView(
-      bufferViews[property.stringOffsetBufferView],
-      offsetType,
-      componentCount + 1
-    );
-  }
-
-  if (hasStrings || hasBooleans) {
-    // STRING and BOOLEAN types need to be parsed differently than other types
-    valueType = MetadataType.UINT8;
-  }
-
-  var valueCount;
-  if (hasStrings) {
-    valueCount = 0;
-    for (i = 0; i < componentCount; ++i) {
-      valueCount += stringOffsets.get(i + 1) - stringOffsets.get(i);
-    }
-  } else if (hasBooleans) {
-    valueCount = Math.ceil(componentCount / 8);
-  } else {
-    valueCount = componentCount;
-  }
-
-  var values = new BufferView(
-    bufferViews[property.bufferView],
-    valueType,
-    valueCount
-  );
-
-  return new MetadataTableProperty(
-    arrayOffsets,
-    stringOffsets,
-    values,
-    classProperty,
-    count
-  );
-}
-
-function getString(index, values, stringOffsets) {
-  var stringByteOffset = stringOffsets.get(index);
-  var stringByteLength = stringOffsets.get(index + 1) - stringByteOffset;
-  return getStringFromTypedArray(
-    values.typedArray,
-    stringByteOffset,
-    stringByteLength
-  );
-}
-
-function getBoolean(index, values) {
-  // byteIndex is floor(index / 8)
-  var byteIndex = index >> 3;
-  var bitIndex = index % 8;
-  return ((values.typedArray[byteIndex] >> bitIndex) & 1) === 1;
-}
-
-function setBoolean(index, values, value) {
-  // byteIndex is floor(index / 8)
-  var byteIndex = index >> 3;
-  var bitIndex = index % 8;
-
-  if (value) {
-    values.typedArray[byteIndex] |= 1 << bitIndex;
-  } else {
-    values.typedArray[byteIndex] &= ~(1 << bitIndex);
-  }
-}
-
-function getInt64NumberFallback(index, values) {
-  var dataView = values.dataView;
-  var byteOffset = index * 8;
-  var value = 0;
-  var isNegative = (dataView.getUint8(byteOffset + 7) & 0x80) > 0;
-  var carrying = true;
-  for (var i = 0; i < 8; ++i) {
-    var byte = dataView.getUint8(byteOffset + i);
-    if (isNegative) {
-      if (carrying) {
-        if (byte !== 0x00) {
-          byte = ~(byte - 1) & 0xff;
-          carrying = false;
-        }
-      } else {
-        byte = ~byte & 0xff;
-      }
-    }
-    value += byte * Math.pow(256, i);
-  }
-  if (isNegative) {
-    value = -value;
-  }
-  return value;
-}
-
-function getInt64BigIntFallback(index, values) {
-  var dataView = values.dataView;
-  var byteOffset = index * 8;
-  var value = BigInt(0); // eslint-disable-line
-  var isNegative = (dataView.getUint8(byteOffset + 7) & 0x80) > 0;
-  var carrying = true;
-  for (var i = 0; i < 8; ++i) {
-    var byte = dataView.getUint8(byteOffset + i);
-    if (isNegative) {
-      if (carrying) {
-        if (byte !== 0x00) {
-          byte = ~(byte - 1) & 0xff;
-          carrying = false;
-        }
-      } else {
-        byte = ~byte & 0xff;
-      }
-    }
-    value += BigInt(byte) * (BigInt(1) << BigInt(i * 8)); // eslint-disable-line
-  }
-  if (isNegative) {
-    value = -value;
-  }
-  return value;
-}
-
-function getUint64NumberFallback(index, values) {
-  var dataView = values.dataView;
-  var byteOffset = index * 8;
-
-  // Split 64-bit number into two 32-bit (4-byte) parts
-  var left = dataView.getUint32(byteOffset, true);
-  var right = dataView.getUint32(byteOffset + 4, true);
-
-  // Combine the two 32-bit values
-  var value = left + 4294967296 * right;
-
-  return value;
-}
-
-function getUint64BigIntFallback(index, values) {
-  var dataView = values.dataView;
-  var byteOffset = index * 8;
-
-  // Split 64-bit number into two 32-bit (4-byte) parts
-  var left = BigInt(dataView.getUint32(byteOffset, true)); // eslint-disable-line
-  var right = BigInt(dataView.getUint32(byteOffset + 4, true)); // eslint-disable-line
-
-  // Combine the two 32-bit values
-  var value = left + BigInt(4294967296) * right; // eslint-disable-line
-
-  return value;
-}
-
-function getComponentDatatype(type) {
-  switch (type) {
-    case MetadataType.INT8:
-      return ComponentDatatype.BYTE;
-    case MetadataType.UINT8:
-      return ComponentDatatype.UNSIGNED_BYTE;
-    case MetadataType.INT16:
-      return ComponentDatatype.SHORT;
-    case MetadataType.UINT16:
-      return ComponentDatatype.UNSIGNED_SHORT;
-    case MetadataType.INT32:
-      return ComponentDatatype.INT;
-    case MetadataType.UINT32:
-      return ComponentDatatype.UNSIGNED_INT;
-    case MetadataType.FLOAT32:
-      return ComponentDatatype.FLOAT;
-    case MetadataType.FLOAT64:
-      return ComponentDatatype.DOUBLE;
-  }
-}
-
-function BufferView(bufferView, type, length) {
-  var that = this;
-
-  var typedArray;
-  var getFunction;
-  var setFunction;
-
-  if (type === MetadataType.INT64) {
-    if (!FeatureDetection.supportsBigInt()) {
-      oneTimeWarning(
-        "INT64 type is not fully supported on this platform. Values greater than 2^53 - 1 or less than -(2^53 - 1) may lose precision when read."
-      );
-      typedArray = new Uint8Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length * 8
-      );
-      getFunction = function (index) {
-        return getInt64NumberFallback(index, that);
-      };
-    } else if (!FeatureDetection.supportsBigInt64Array()) {
-      typedArray = new Uint8Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length * 8
-      );
-      getFunction = function (index) {
-        return getInt64BigIntFallback(index, that);
-      };
-    } else {
-      // eslint-disable-next-line
-      typedArray = new BigInt64Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length
-      );
-      setFunction = function (index, value) {
-        // Convert the number to a BigInt before setting the value in the typed array
-        that.typedArray[index] = BigInt(value); // eslint-disable-line
-      };
-    }
-  } else if (type === MetadataType.UINT64) {
-    if (!FeatureDetection.supportsBigInt()) {
-      oneTimeWarning(
-        "UINT64 type is not fully supported on this platform. Values greater than 2^53 - 1 may lose precision when read."
-      );
-      typedArray = new Uint8Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length * 8
-      );
-      getFunction = function (index) {
-        return getUint64NumberFallback(index, that);
-      };
-    } else if (!FeatureDetection.supportsBigUint64Array()) {
-      typedArray = new Uint8Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length * 8
-      );
-      getFunction = function (index) {
-        return getUint64BigIntFallback(index, that);
-      };
-    } else {
-      // eslint-disable-next-line
-      typedArray = new BigUint64Array(
-        bufferView.buffer,
-        bufferView.byteOffset,
-        length
-      );
-      setFunction = function (index, value) {
-        // Convert the number to a BigInt before setting the value in the typed array
-        that.typedArray[index] = BigInt(value); // eslint-disable-line
-      };
-    }
-  } else {
-    var componentDatatype = getComponentDatatype(type);
-    typedArray = ComponentDatatype.createArrayBufferView(
-      componentDatatype,
-      bufferView.buffer,
-      bufferView.byteOffset,
-      length
-    );
-    setFunction = function (index, value) {
-      that.typedArray[index] = value;
-    };
-  }
-
-  if (!defined(getFunction)) {
-    getFunction = function (index) {
-      return that.typedArray[index];
-    };
-  }
-
-  this.typedArray = typedArray;
-  this.dataView = new DataView(typedArray.buffer, typedArray.byteOffset);
-  this.get = getFunction;
-  this.set = setFunction;
-}
-
-function MetadataTableProperty(
-  arrayOffsets,
-  stringOffsets,
-  values,
-  classProperty,
-  count
-) {
-  var that = this;
-
-  var getValueFunction;
-  var setValueFunction;
-
-  var valueType = classProperty.valueType;
-  var enumType = classProperty.enumType;
-
-  if (valueType === MetadataType.STRING) {
-    getValueFunction = function (index) {
-      return getString(index, that.values, that.stringOffsets);
-    };
-  } else if (valueType === MetadataType.BOOLEAN) {
-    getValueFunction = function (index) {
-      return getBoolean(index, that.values);
-    };
-    setValueFunction = function (index, value) {
-      setBoolean(index, that.values, value);
-    };
-  } else if (defined(enumType)) {
-    getValueFunction = function (index) {
-      var integer = that.values.get(index);
-      return enumType.namesByValue[integer];
-    };
-    setValueFunction = function (index, value) {
-      var integer = enumType.valuesByName[value];
-      that.values.set(index, integer);
-    };
-  } else {
-    getValueFunction = function (index) {
-      return that.values.get(index);
-    };
-    setValueFunction = function (index, value) {
-      that.values.set(index, value);
-    };
-  }
-
-  this.arrayOffsets = arrayOffsets;
-  this.stringOffsets = stringOffsets;
-  this.values = values;
-  this.classProperty = classProperty;
-  this.count = count;
-  this.getValue = getValueFunction;
-  this.setValue = setValueFunction;
-  this.unpackedValues = undefined;
-}
-
-MetadataTableProperty.prototype.get = function (index) {
-  if (requiresUnpackForGet(this)) {
-    unpackProperty(this);
-  }
-
-  var classProperty = this.classProperty;
-
-  if (defined(this.unpackedValues)) {
-    var value = this.unpackedValues[index];
-    if (classProperty.type === MetadataType.ARRAY) {
-      return value.slice(); // clone
-    }
-    return value;
-  }
-
-  if (classProperty.type !== MetadataType.ARRAY) {
-    return this.getValue(index);
-  }
-
-  var offset;
-  var length;
-
-  var componentCount = classProperty.componentCount;
-  if (defined(componentCount)) {
-    offset = index * componentCount;
-    length = componentCount;
-  } else {
-    offset = this.arrayOffsets.get(index);
-    length = this.arrayOffsets.get(index + 1) - offset;
-  }
-
-  var values = new Array(length);
-  for (var i = 0; i < length; ++i) {
-    values[i] = this.getValue(offset + i);
-  }
-
-  return values;
-};
-
-MetadataTableProperty.prototype.set = function (index, value) {
-  if (requiresUnpackForSet(this, index, value)) {
-    unpackProperty(this);
-  }
-
-  var classProperty = this.classProperty;
-
-  if (defined(this.unpackedValues)) {
-    if (classProperty.type === MetadataType.ARRAY) {
-      value = value.slice(); // clone
-    }
-    this.unpackedValues[index] = value;
-    return;
-  }
-
-  // Values are unpacked if the length of a variable-size array changes or the
-  // property has strings. No need to handle these cases below.
-
-  if (classProperty.type !== MetadataType.ARRAY) {
-    this.setValue(index, value);
-    return;
-  }
-
-  var offset;
-  var length;
-
-  var componentCount = classProperty.componentCount;
-  if (defined(componentCount)) {
-    offset = index * componentCount;
-    length = componentCount;
-  } else {
-    offset = this.arrayOffsets.get(index);
-    length = this.arrayOffsets.get(index + 1) - offset;
-  }
-
-  for (var i = 0; i < length; ++i) {
-    this.setValue(offset + i, value[i]);
-  }
-};
-
-function requiresUnpackForGet(property) {
-  if (defined(property.unpackedValues)) {
-    return false;
-  }
-
-  var valueType = property.classProperty.valueType;
-
-  if (valueType === MetadataType.STRING) {
-    // Unpack since UTF-8 decoding is expensive
-    return true;
-  }
-
-  if (
-    valueType === MetadataType.INT64 &&
-    !FeatureDetection.supportsBigInt64Array()
-  ) {
-    // Unpack since the fallback INT64 getters are expensive
-    return true;
-  }
-
-  if (
-    valueType === MetadataType.UINT64 &&
-    !FeatureDetection.supportsBigUint64Array()
-  ) {
-    // Unpack since the fallback UINT64 getters are expensive
-    return true;
-  }
-
-  return false;
-}
-
-function requiresUnpackForSet(property, index, value) {
-  if (requiresUnpackForGet(property)) {
-    return true;
-  }
-
-  var arrayOffsets = property.arrayOffsets;
-  if (defined(arrayOffsets)) {
-    // Unpacking is required if a variable-size array changes length since it
-    // would be expensive to repack the binary data
-    var oldLength = arrayOffsets.get(index + 1) - arrayOffsets.get(index);
-    var newLength = value.length;
-    if (oldLength !== newLength) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-function unpackProperty(property) {
-  property.unpackedValues = unpackValues(property);
-
-  // Free memory
-  property.arrayOffsets = undefined;
-  property.stringOffsets = undefined;
-  property.values = undefined;
-}
-
-function unpackValues(property) {
-  var i;
-  var count = property.count;
-  var unpackedValues = new Array(count);
-
-  var classProperty = property.classProperty;
-  if (classProperty.type !== MetadataType.ARRAY) {
-    for (i = 0; i < count; ++i) {
-      unpackedValues[i] = property.getValue(i);
-    }
-    return unpackedValues;
-  }
-
-  var j;
-  var offset;
-  var arrayValues;
-
-  var componentCount = classProperty.componentCount;
-  if (defined(componentCount)) {
-    for (i = 0; i < count; ++i) {
-      arrayValues = new Array(componentCount);
-      unpackedValues[i] = arrayValues;
-      offset = i * componentCount;
-      for (j = 0; j < componentCount; ++j) {
-        arrayValues[j] = property.getValue(offset + j);
-      }
-    }
-    return unpackedValues;
-  }
-
-  for (i = 0; i < count; ++i) {
-    offset = property.arrayOffsets.get(i);
-    var length = property.arrayOffsets.get(i + 1) - offset;
-    arrayValues = new Array(length);
-    unpackedValues[i] = arrayValues;
-    for (j = 0; j < length; ++j) {
-      arrayValues[j] = property.getValue(offset + j);
-    }
-  }
-
-  return unpackedValues;
 }
 
 export default MetadataTable;
