+++ Specs/Scene/MetadataClassPropertySpec.js
@@ -1,3 +1,4 @@
+import { FeatureDetection } from "../../Source/Cesium.js";
 import { MetadataClassProperty } from "../../Source/Cesium.js";
 import { MetadataEnum } from "../../Source/Cesium.js";
 import { MetadataType } from "../../Source/Cesium.js";
@@ -124,5 +125,198 @@ describe("Scene/MetadataClassProperty", function () {
         id: "propertyId",
       });
     }).toThrowDeveloperError();
+  });
+
+  it("validate returns undefined if the value is valid", function () {
+    var property = new MetadataClassProperty({
+      id: "position",
+      property: {
+        type: "ARRAY",
+        componentType: "FLOAT32",
+        componentCount: 3,
+      },
+    });
+
+    expect(property.validate([1.0, 2.0, 3.0])).toBeUndefined();
+  });
+
+  it("validate returns error message if type is ARRAY and value is not an array", function () {
+    var property = new MetadataClassProperty({
+      id: "position",
+      property: {
+        type: "ARRAY",
+        componentType: "FLOAT32",
+        componentCount: 3,
+      },
+    });
+
+    expect(property.validate(8.0)).toBe("value 8 does not match type ARRAY");
+  });
+
+  it("validate returns error message if type is ARRAY and the array length does not match componentCount", function () {
+    var property = new MetadataClassProperty({
+      id: "position",
+      property: {
+        type: "ARRAY",
+        componentType: "FLOAT32",
+        componentCount: 3,
+      },
+    });
+
+    expect(property.validate([1.0, 2.0])).toBe(
+      "Array length does not match componentCount"
+    );
+  });
+
+  it("validate returns error message if enum name is invalid", function () {
+    var myEnum = new MetadataEnum({
+      id: "myEnum",
+      enum: {
+        values: [
+          {
+            value: 0,
+            name: "ValueA",
+          },
+          {
+            value: 1,
+            name: "ValueB",
+          },
+          {
+            value: -999,
+            name: "Other",
+          },
+        ],
+      },
+    });
+
+    var property = new MetadataClassProperty({
+      id: "myEnum",
+      property: {
+        type: "ENUM",
+        enumType: "myEnum",
+      },
+      enums: {
+        myEnum: myEnum,
+      },
+    });
+
+    expect(property.validate("INVALID")).toBe(
+      "value INVALID is not a valid enum name for myEnum"
+    );
+    expect(property.validate(0)).toBe(
+      "value 0 is not a valid enum name for myEnum"
+    );
+  });
+
+  it("validate returns error message if value does not match the type", function () {
+    var types = [
+      "INT8",
+      "UINT8",
+      "INT16",
+      "UINT16",
+      "INT32",
+      "UINT32",
+      "INT64",
+      "UINT64",
+      "FLOAT32",
+      "FLOAT64",
+      "BOOLEAN",
+      "STRING",
+    ];
+
+    for (var i = 0; i < types.length; ++i) {
+      var property = new MetadataClassProperty({
+        id: "property",
+        property: {
+          type: types[i],
+        },
+      });
+      expect(property.validate({})).toBe(
+        "value [object Object] does not match type " + types[i]
+      );
+    }
+  });
+
+  it("validate returns error message if value is out of range", function () {
+    if (!FeatureDetection.supportsBigInt()) {
+      return;
+    }
+
+    var outOfRangeValues = {
+      INT8: [-129, 128],
+      UINT8: [-1, 256],
+      INT16: [-32769, 32768],
+      UINT16: [-1, 65536],
+      INT32: [-2147483649, 2147483648],
+      UINT32: [-1, 4294967296],
+      INT64: [
+        BigInt("-9223372036854775809"), // eslint-disable-line
+        BigInt("9223372036854775808"), // eslint-disable-line
+      ],
+      UINT64: [
+        BigInt(-1), // eslint-disable-line
+        BigInt("18446744073709551616"), // eslint-disable-line
+      ],
+      FLOAT32: [-Number.MAX_VALUE, Number.MAX_VALUE],
+    };
+
+    for (var type in outOfRangeValues) {
+      if (outOfRangeValues.hasOwnProperty(type)) {
+        var values = outOfRangeValues[type];
+        var property = new MetadataClassProperty({
+          id: "property",
+          property: {
+            type: type,
+          },
+        });
+        for (var i = 0; i < values.length; ++i) {
+          expect(property.validate(values[i])).toBe(
+            "value " + values[i] + " is out of range for type " + type
+          );
+        }
+      }
+    }
+  });
+
+  it("validate returns error message if component value is out of range", function () {
+    if (!FeatureDetection.supportsBigInt()) {
+      return;
+    }
+
+    var outOfRangeValues = {
+      INT8: [-129, 128],
+      UINT8: [-1, 256],
+      INT16: [-32769, 32768],
+      UINT16: [-1, 65536],
+      INT32: [-2147483649, 2147483648],
+      UINT32: [-1, 4294967296],
+      INT64: [
+        BigInt("-9223372036854775809"), // eslint-disable-line
+        BigInt("9223372036854775808"), // eslint-disable-line
+      ],
+      UINT64: [
+        BigInt(-1), // eslint-disable-line
+        BigInt("18446744073709551616"), // eslint-disable-line
+      ],
+      FLOAT32: [-Number.MAX_VALUE, Number.MAX_VALUE],
+    };
+
+    for (var componentType in outOfRangeValues) {
+      if (outOfRangeValues.hasOwnProperty(componentType)) {
+        var values = outOfRangeValues[componentType];
+        var property = new MetadataClassProperty({
+          id: "property",
+          property: {
+            type: "ARRAY",
+            componentType: componentType,
+          },
+        });
+        for (var i = 0; i < values.length; ++i) {
+          expect(property.validate(values)).toBe(
+            "value " + values[0] + " is out of range for type " + componentType
+          );
+        }
+      }
+    }
   });
 });
