+++ Specs/Scene/MetadataTableSpec.js
@@ -1636,7 +1636,7 @@ describe("Scene/MetadataTable", function () {
     }
   });
 
-  it("setProperty throws if type is ARRAY and value is not an array", function () {
+  it("setProperty throws if value doesn't conform to the class property", function () {
     var properties = {
       position: {
         type: "ARRAY",
@@ -1655,251 +1655,6 @@ describe("Scene/MetadataTable", function () {
     expect(function () {
       metadataTable.setProperty(0, "position", 8.0);
     }).toThrowDeveloperError();
-  });
-
-  it("setProperty throws if type is ARRAY and the array length does not match componentCount", function () {
-    var properties = {
-      position: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-        componentCount: 3,
-      },
-    };
-    var propertyValues = {
-      position: [[1.0, 2.0, 3.0]],
-    };
-    var metadataTable = createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    expect(function () {
-      metadataTable.setProperty(0, "position", [1.0, 2.0]);
-    }).toThrowDeveloperError();
-  });
-
-  it("setProperty throws if enum name is invalid", function () {
-    var properties = {
-      type: {
-        type: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-    var propertyValues = {
-      type: ["Other"],
-    };
-    var metadataTable = createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    expect(function () {
-      metadataTable.setProperty(0, "type", "INVALID");
-    }).toThrowDeveloperError();
-  });
-
-  it("setProperty throws if value does not match the type", function () {
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-      },
-      propertyUint8: {
-        type: "UINT8",
-      },
-      propertyInt16: {
-        type: "INT16",
-      },
-      propertyUint16: {
-        type: "UINT16",
-      },
-      propertyInt32: {
-        type: "INT32",
-      },
-      propertyUint32: {
-        type: "UINT32",
-      },
-      propertyInt64: {
-        type: "INT64",
-      },
-      propertyUint64: {
-        type: "UINT64",
-      },
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "BOOLEAN",
-      },
-      propertyString: {
-        type: "STRING",
-      },
-      propertyEnum: {
-        type: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [10],
-      propertyUint8: [10],
-      propertyInt16: [10],
-      propertyUint16: [10],
-      propertyInt32: [10],
-      propertyUint32: [10],
-      propertyInt64: [BigInt(10)], // eslint-disable-line
-      propertyUint64: [BigInt(10)], // eslint-disable-line
-      propertyFloat32: [10.0],
-      propertyFloat64: [10.0],
-      propertyBoolean: [true],
-      propertyString: ["a"],
-      propertyEnum: ["Other"],
-    };
-
-    var metadataTable = createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    function getSetFunction(propertyId) {
-      return function () {
-        expect(function () {
-          metadataTable.setProperty(0, propertyId, {});
-        }).toThrowDeveloperError();
-      };
-    }
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        getSetFunction(propertyId)();
-      }
-    }
-  });
-
-  it("setProperty throws if value is out of range", function () {
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-      },
-      propertyUint8: {
-        type: "UINT8",
-      },
-      propertyInt16: {
-        type: "INT16",
-      },
-      propertyUint16: {
-        type: "UINT16",
-      },
-      propertyInt32: {
-        type: "INT32",
-      },
-      propertyUint32: {
-        type: "UINT32",
-      },
-      propertyInt64: {
-        type: "INT64",
-      },
-      propertyUint64: {
-        type: "UINT64",
-      },
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [10],
-      propertyUint8: [10],
-      propertyInt16: [10],
-      propertyUint16: [10],
-      propertyInt32: [10],
-      propertyUint32: [10],
-      propertyInt64: [BigInt(10)], // eslint-disable-line
-      propertyUint64: [BigInt(10)], // eslint-disable-line
-      propertyFloat32: [10.0],
-    };
-
-    var outOfRangeValues = {
-      propertyInt8: [-129, 128],
-      propertyUint8: [-1, 256],
-      propertyInt16: [-32769, 32768],
-      propertyUint16: [-1, 65536],
-      propertyInt32: [-2147483649, 2147483648],
-      propertyUint32: [-1, 4294967296],
-      propertyInt64: [
-        BigInt("-9223372036854775809"), // eslint-disable-line
-        BigInt("9223372036854775808"), // eslint-disable-line
-      ],
-      propertyUint64: [BigInt(-1), BigInt("18446744073709551616")], // eslint-disable-line
-      propertyFloat32: [-Number.MAX_VALUE, Number.MAX_VALUE],
-    };
-
-    var metadataTable = createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    function getSetFunction(propertyId, value) {
-      return function () {
-        expect(function () {
-          metadataTable.setProperty(0, propertyId, value);
-        }).toThrowDeveloperError();
-      };
-    }
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var valuesToTest = outOfRangeValues[propertyId];
-        for (var i = 0; i < valuesToTest.length; ++i) {
-          getSetFunction(propertyId, valuesToTest[i])();
-        }
-      }
-    }
-  });
-
-  it("setProperty throws if component value is out of range", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-        componentCount: 2,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [[0]],
-      propertyUint8: [[0, 0]],
-    };
-
-    var outOfRangeValues = {
-      propertyInt8: [-129, 128, 129],
-      propertyUint8: [-1, 256],
-    };
-
-    var metadataTable = createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    function getSetFunction(propertyId, value) {
-      return function () {
-        expect(function () {
-          metadataTable.setProperty(0, propertyId, value);
-        }).toThrowDeveloperError();
-      };
-    }
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        getSetFunction(propertyId, outOfRangeValues[propertyId])();
-      }
-    }
   });
 
   it("setProperty sets Infinity for FLOAT32 and FLOAT64", function () {
